+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/header/header';
+import Header from '../../components/headerComponent/headerComponent';
 import PlaceCard from '../../components/place-card/place-card';
 
 function FavoritesPage (): JSX.Element{
